# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks/{taskId}/video-links:
    get:
      operationId: listVideoLinks
      tags: [tasks]
      summary: List video links for a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Video links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoLink"
    post:
      operationId: addVideoLink
      tags: [tasks]
      summary: Add a video link to a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoLinkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoLink"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /video-links/{videoLinkId}:
    patch:
      operationId: updateVideoLink
      tags: [tasks]
      summary: Update a video link
      parameters:
        - $ref: "#/components/parameters/VideoLinkId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoLinkUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoLink"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteVideoLink
      tags: [tasks]
      summary: Delete a video link
      parameters:
        - $ref: "#/components/parameters/VideoLinkId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: integer
...
    VideoLinkId:
      name: videoLinkId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    VideoLink:
      type: object
      required: [id, taskId, url, position, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        url:
          type: string
        label:
          type: ["string", "null"]
        position:
          type: integer
        createdAt:
          type: string
    VideoLinkInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          minLength: 1
        label:
          type: ["string", "null"]
    VideoLinkUpdate:
      type: object
      properties:
        url:
          type: string
          minLength: 1
        label:
          type: ["string", "null"]
...